feat: restructure module and add type stubs for improved clarity and functionality
