+++ python/pyproject.toml
@@ -15,6 +15,8 @@ build-backend = "maturin"
 [tool.maturin]
 manifest-path = "../crates/timeseries-table-python/Cargo.toml"
 bindings = "pyo3"
+python-source = "src"
+module-name = "timeseries_table_format._dev"
 
 [dependency-groups]
 dev = [
