+++ python/src/timeseries_table_format/__init__.py
@@ -1,0 +1,3 @@
+from ._dev import *
+__doc__ = _dev.__doc__
+if hasattr(_dev, "__all__"): __all__ = _dev.__all__
